+++ iam-policy-secretsmanager.tf
@@ -4,9 +4,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  */
 
-resource "aws_iam_role_policy" "secretsmanager_for_ecs_readonly" {
+resource "aws_iam_role_policy" "secretsmanager_for_hms_readonly" {
   name = "secretsmanager"
-  role = "${aws_iam_role.apiary_task_readonly.id}"
+  role = "${aws_iam_role.apiary_hms_readonly.id}"
 
   policy = <<EOF
 {
@@ -14,15 +14,15 @@ resource "aws_iam_role_policy" "secretsmanager_for_ecs_readonly" {
     "Statement": {
         "Effect": "Allow",
         "Action": "secretsmanager:GetSecretValue",
-        "Resource": [ "${join("\",\"",concat(data.aws_secretsmanager_secret.db_ro_user.*.arn,data.aws_secretsmanager_secret.ldap_user.*.arn,data.aws_secretsmanager_secret.ranger_audit.*.arn))}" ]
+        "Resource": [ "${join("\",\"", concat(data.aws_secretsmanager_secret.db_ro_user.*.arn, data.aws_secretsmanager_secret.ldap_user.*.arn, data.aws_secretsmanager_secret.ranger_audit.*.arn))}" ]
     }
 }
 EOF
 }
 
-resource "aws_iam_role_policy" "secretsmanager_for_ecs_task_readwrite" {
+resource "aws_iam_role_policy" "secretsmanager_for_hms_readwrite" {
   name = "secretsmanager"
-  role = "${aws_iam_role.apiary_task_readwrite.id}"
+  role = "${aws_iam_role.apiary_hms_readwrite.id}"
 
   policy = <<EOF
 {
@@ -30,14 +30,14 @@ resource "aws_iam_role_policy" "secretsmanager_for_ecs_task_readwrite" {
     "Statement": {
         "Effect": "Allow",
         "Action": "secretsmanager:GetSecretValue",
-        "Resource": [ "${join("\",\"",concat(data.aws_secretsmanager_secret.db_rw_user.*.arn,data.aws_secretsmanager_secret.ldap_user.*.arn,data.aws_secretsmanager_secret.ranger_audit.*.arn))}" ]
+        "Resource": [ "${join("\",\"", concat(data.aws_secretsmanager_secret.db_rw_user.*.arn, data.aws_secretsmanager_secret.ldap_user.*.arn, data.aws_secretsmanager_secret.ranger_audit.*.arn))}" ]
     }
 }
 EOF
 }
 
 resource "aws_iam_role_policy" "secretsmanager_for_ecs_task_exec" {
-  count = "${var.docker_registry_auth_secret_name == "" ? 0 : 1}"
+  count = "${var.hms_instance_type == "ecs" && var.docker_registry_auth_secret_name != "" ? 1 : 0}"
   name  = "secretsmanager-exec"
   role  = "${aws_iam_role.apiary_task_exec.id}"
 
@@ -47,7 +47,7 @@ resource "aws_iam_role_policy" "secretsmanager_for_ecs_task_exec" {
     "Statement": {
         "Effect": "Allow",
         "Action": "secretsmanager:GetSecretValue",
-        "Resource": [ "${join("\",\"",concat(data.aws_secretsmanager_secret.docker_registry.*.arn))}" ]
+        "Resource": [ "${join("\",\"", concat(data.aws_secretsmanager_secret.docker_registry.*.arn))}" ]
     }
 }
 EOF